# apps/ios/Packages/DenAPI/Sources/DenAPI/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Den API",
    "description": "M3 REST API. Cookie-authenticated writes require Origin and X-CSRF-Token. CLI and agents use bearer authentication. Entity IDs are ULIDs; background image IDs are opaque SHA-256 content IDs.",
    "license": {
      "name": "AGPL-3.0-only",
      "identifier": "AGPL-3.0-only"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/channels/{id}/threads": {
      "get": {
        "tags": [
          "thread_read"
        ],
        "operationId": "get__channels__id__threads",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "resolved",
            "in": "query",
            "description": "Absent lists every thread; false lists the open strip; true lists resolved history.",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "unread_only",
            "in": "query",
            "description": "True lists only threads contributing unread activity to this caller's channel\ntotal. Absent or false applies no unread filter.",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "before",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ThreadView"
                  }
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "default": {
            "description": "API error; invalid input, authentication, permissions, conflict, throttling or storage failure",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
...
    "/threads/{id}": {
      "get": {
        "tags": [
          "thread_read"
        ],
        "operationId": "get__threads__id_",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ThreadView"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "default": {
            "description": "API error; invalid input, authentication, permissions, conflict, throttling or storage failure",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "threads"
        ],
        "operationId": "patch__threads__id_",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateThread"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ThreadSummary"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "default": {
            "description": "API error; invalid input, authentication, permissions, conflict, throttling or storage failure",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/threads/{id}/follow": {
      "put": {
        "tags": [
          "thread_read"
        ],
        "operationId": "put__threads__id__follow",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FollowThread"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ThreadReadState"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "default": {
            "description": "API error; invalid input, authentication, permissions, conflict, throttling or storage failure",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
...
    "/threads/{id}/read": {
      "put": {
        "tags": [
          "thread_read"
        ],
        "operationId": "put__threads__id__read",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MarkThreadRead"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ThreadReadState"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "default": {
            "description": "API error; invalid input, authentication, permissions, conflict, throttling or storage failure",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApiError": {
        "type": "object",
        "required": [
          "error",
          "message"
        ],
        "properties": {
          "error": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        }
      },
...
      "FollowThread": {
        "type": "object",
        "required": [
          "following"
        ],
        "properties": {
          "following": {
            "type": "boolean"
          }
        }
      },
...
      "MarkThreadRead": {
        "type": "object",
        "description": "Advances one thread's own read position. A thread read never moves the main\nconversation's position, and reading does not by itself follow the thread.",
        "required": [
          "message_id"
        ],
        "properties": {
          "message_id": {
            "$ref": "#/components/schemas/String"
          }
        }
      },
...
      "String": {
        "type": "string"
      },
...
      "ThreadReadState": {
        "type": "object",
        "description": "One user's position in one thread. Following changes unread accounting, never access.",
        "required": [
          "thread_id",
          "channel_id",
          "following",
          "unread_count",
          "mention_count",
          "notification_count"
        ],
        "properties": {
          "channel_id": {
            "$ref": "#/components/schemas/String"
          },
          "following": {
            "type": "boolean"
          },
          "last_read_id": {
            "type": [
              "string",
              "null"
            ]
          },
          "mention_count": {
            "type": "integer",
            "format": "int64"
          },
          "notification_count": {
            "type": "integer",
            "format": "int64",
            "description": "Unread replies eligible under current preferences, excluding own messages."
          },
          "thread_id": {
            "$ref": "#/components/schemas/String"
          },
          "unread_count": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "ThreadSummary": {
        "type": "object",
        "description": "Shared thread metadata. Deliberately free of per-user state so one payload can\nbe broadcast to every member who can see the parent channel.",
        "required": [
          "id",
          "channel_id",
          "root_message_id",
          "title",
          "created_by",
          "created_at",
          "reply_count",
          "last_activity_at"
        ],
        "properties": {
          "channel_id": {
            "$ref": "#/components/schemas/String"
          },
          "created_at": {
            "type": "string"
          },
          "created_by": {
            "$ref": "#/components/schemas/String"
          },
          "id": {
            "$ref": "#/components/schemas/String"
          },
          "last_activity_at": {
            "type": "string"
          },
          "last_reply_id": {
            "type": [
              "string",
              "null"
            ]
          },
          "reply_count": {
            "type": "integer",
            "format": "int64"
          },
          "resolved_at": {
            "type": [
              "string",
              "null"
            ],
            "description": "Both resolution fields are set together or not at all."
          },
          "resolved_by": {
            "type": [
              "string",
              "null"
            ]
          },
          "root_message_id": {
            "$ref": "#/components/schemas/String",
            "description": "The message the conversation hangs off. It stays in the main conversation."
          },
          "title": {
            "type": "string",
            "description": "A snapshot inferred at creation. Editing the root never rewrites it."
          }
        }
      },
      "ThreadView": {
        "type": "object",
        "required": [
          "thread",
          "read_state"
        ],
        "properties": {
          "read_state": {
            "$ref": "#/components/schemas/ThreadReadState"
          },
          "thread": {
            "$ref": "#/components/schemas/ThreadSummary"
          }
        }
      },
...
      "UpdateThread": {
        "type": "object",
        "description": "Rename, resolve or reopen. An omitted field is left unchanged.",
        "properties": {
          "resolved": {
            "type": [
              "boolean",
              "null"
            ]
          },
          "title": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      },
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "den_session"
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    },
    {
      "sessionCookie": []
    }
  ]
}